Account for ktlint.
If ktlint was loaded and added to checkers already, gradle will take
precedence over ktlint so make sure to set ktlint to go next.

If ktlint isn't loaded yet (or ever), set it so that when it does it
load (it will now take precedence over gradle), so gradle will go
after ktlint.